one-pager/orchester/requirements.txt
functions-framework==3.*
requests
markdown
pdfkit
google-auth
google-cloud-storage
flask
reportlab
weasyprint
pyshorteners